# src/main/java/org/example/datahub/api/UsersApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.1.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package org.example.datahub.api;

import org.example.datahub.model.UserDetailResponseDTO;
import org.example.datahub.model.UserSignUpRequestDTO;
import org.example.datahub.model.UserSignUpResponseDTO;
import org.example.datahub.model.UserVerifyRoleRequestDTO;
import org.example.datahub.model.UserVerifyRoleResponseDTO;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2025-12-04T09:26:02.924376840+08:00[Asia/Shanghai]")
@Validated
@Tag(name = "users", description = "the users API")
public interface UsersApi {

    /**
     * GET /users/{user_id} : Get user detail
     * Query user detail.
     *
     * @param userId  (required)
     * @return Successful response (status code 200)
     */
    @Operation(
        operationId = "userDetail",
        summary = "Get user detail",
        description = "Query user detail.",
        responses = {
            @ApiResponse(responseCode = "200", description = "Successful response", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = UserDetailResponseDTO.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/users/{user_id}",
        produces = { "application/json" }
    )
    
    ResponseEntity<UserDetailResponseDTO> userDetail(
        @Parameter(name = "user_id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("user_id") Long userId
    );


    /**
     * POST /users/signup : User signup
     * Create a new user.
     *
     * @param userSignUpRequestDTO  (required)
     * @return Successful response (status code 200)
     */
    @Operation(
        operationId = "userSignUp",
        summary = "User signup",
        description = "Create a new user.",
        responses = {
            @ApiResponse(responseCode = "200", description = "Successful response", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = UserSignUpResponseDTO.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/users/signup",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    ResponseEntity<UserSignUpResponseDTO> userSignUp(
        @Parameter(name = "UserSignUpRequestDTO", description = "", required = true) @Valid @RequestBody UserSignUpRequestDTO userSignUpRequestDTO
    );


    /**
     * POST /users/{user_id}/verify-role : Verify user role
     * Verify and update user role.
     *
     * @param userId  (required)
     * @param userVerifyRoleRequestDTO  (required)
     * @return Successful response (status code 200)
     */
    @Operation(
        operationId = "userVerifyRole",
        summary = "Verify user role",
        description = "Verify and update user role.",
        responses = {
            @ApiResponse(responseCode = "200", description = "Successful response", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = UserVerifyRoleResponseDTO.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/users/{user_id}/verify-role",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    ResponseEntity<UserVerifyRoleResponseDTO> userVerifyRole(
        @Parameter(name = "user_id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("user_id") Long userId,
        @Parameter(name = "UserVerifyRoleRequestDTO", description = "", required = true) @Valid @RequestBody UserVerifyRoleRequestDTO userVerifyRoleRequestDTO
    );

}