# src/main/resources/openapi/api.yml
openapi: 3.0.3
info:
  title: Research Collection System API
  version: 1.0.0
  description: API specification for research task collection system.
servers:
  - url: /api
paths:
...
  /users/signup:
    post:
      operationId: userSignUp
      summary: User signup
      description: Create a new user.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserSignUpRequest"
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserSignUpResponse"

#  /users/signin:
#    post:
#      operationId: userSignIn
#      summary: User signin
#      description: Authenticate user and return JWT.
#      requestBody:
#        required: true
#        content:
#          application/json:
#            schema:
#              $ref: "#/components/schemas/UserSignInRequest"
#      responses:
#        "200":
#          description: Successful response
#          content:
#            application/json:
#              schema:
#                $ref: "#/components/schemas/UserSignInResponse"
#
#  /users/signout:
#    post:
#      operationId: userSignOut
#      summary: User signout
#      description: Sign out user.
#      responses:
#        "200":
#          description: Successful response
#          content:
#            application/json:
#              schema:
#                $ref: "#/components/schemas/SuccessResponse"
  /users/{user_id}:
    get:
      operationId: userDetail
      summary: Get user detail
      description: Query user detail.
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDetailResponse"
  /users/{user_id}/verify-role:
    post:
      operationId: userVerifyRole
      summary: Verify user role
      description: Verify and update user role.
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserVerifyRoleRequest"
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserVerifyRoleResponse"
...
components:
  schemas:
...
    UserSignUpRequest:
      type: object
      properties:
        username: { type: string }
        password: { type: string }
        user_email: { type: string }
        role: { type: string }
      required: [username, password, user_email, role]

#    UserSignInRequest:
#      type: object
#      properties:
#        username: { type: string }
#        password: { type: string }
#      required: [username, password]
    UserVerifyRoleRequest:
      type: object
      properties:
        role: { type: string }
        employee_id: { type: string }
        assistant_name: { type: string }
        email_app_password: { type: string }
      required: [role, employee_id, assistant_name, email_app_password]
    UserSignUpResponse:
      type: object
      properties:
        success: { type: boolean }
        data:
          type: object
          properties:
            user_id: { type: integer, format: int64 }
        message: { type: string }

#    UserSignInResponse:
#      type: object
#      properties:
#        success: { type: boolean }
#        data:
#          type: object
#          properties:
#            token: { type: string }
#            user:
#              type: object
#              properties:
#                user_id: { type: integer, format: int64 }
#                username: { type: string }
#                role: { type: string }
    UserDetailResponse:
      type: object
      properties:
        success: { type: boolean }
        data:
          type: object
          properties:
            user:
              type: object
              properties:
                user_id: { type: integer, format: int64 }
                username: { type: string }
                user_email: { type: string }
                role: { type: string }
                assistant_id: { type: integer, format: int64 }
        message: { type: string }
    UserVerifyRoleResponse:
      type: object
      properties:
        success: { type: boolean }
        data:
          type: object
          properties:
            role: { type: string }
            assistant_id: { type: integer, format: int64 }
        message: { type: string }

    # ==================== TASK ====================
...
    SuccessResponse:
      type: object
      properties:
        success: { type: boolean }
        message: { type: string }
...